weevil-lunar/requirements.txt
numpy>=1.26,<2.0; python_version < '3.13'
numpy>=2.1; python_version >= '3.13'
scipy>=1.11; python_version < '3.13'
scipy>=1.14; python_version >= '3.13'
matplotlib>=3.8
pandas>=2.1
pyyaml>=6.0
pytest>=8.0
pytest-cov>=5.0
ruff>=0.6
black>=24.0
yamllint>=1.35
jsonschema>=4.0
pymarkdownlnt>=0.9
jupyter>=1.0
nbformat>=5.10
